// client/src/app/features/minrps/models/enums/minrps-game-event.enum.ts
export enum MinRpsGameEvent {
  Connected = 'minrps.game.connected',
  Disconnected = 'minrps.game.disconnected',
  Join = 'minrps.game.join',
  Joined = 'minrps.game.joined',
  Leave = 'minrps.game.leave',
  Left = 'minrps.game.left',
  TakeSeat = 'minrps.game.take-seat',
  SelectMove = 'minrps.game.select-move',
  MoveSelected = 'minrps.game.move-selected',
  Play = 'minrps.game.play',
  Played = 'minrps.game.played',
  GameStateUpdate = 'minrps.game.state-update',
}